backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
asyncpg==0.29.0
pydantic==2.9.0
python-jose[cryptography]==3.3.0
httpx==0.27.0
python-dotenv==1.0.1
python-multipart==0.0.12
